database_api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.6.1
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.26.0
python-multipart>=0.0.9
alembic>=1.13.1 